Extract DropzoneText
